Move registration form to its own component
