+++ lunchbreak-frontend/src/components/inputForms/RegistrationForm.js
@@ -1,0 +1,111 @@
+import React from "react";
+import Grid from "@material-ui/core/Grid";
+import Typography from "@material-ui/core/Typography";
+import {Form, Formik} from "formik";
+import {performRegistration} from "../../utils/AuthUtils";
+import * as Yup from "yup";
+import InputTextFieldValidated from "../inputFields/InputTextFieldValidated";
+import ButtonYellowBig from "../buttons/ButtonYellowBig";
+import makeStyles from "@material-ui/core/styles/makeStyles";
+
+const useStyles = makeStyles((theme) => ({
+    nextTopic: {
+        paddingTop: theme.spacing(3)
+    }
+}));
+
+export default function RegistrationForm() {
+    const classes = useStyles();
+
+    function cancel() {
+        window.location.href = "/login";
+    }
+
+    return (
+        <>
+            <Grid item>
+                <Typography variant="h4" color="primary" align="center">
+                    Registrierung
+                </Typography>
+            </Grid>
+            <Formik initialValues={
+                {
+                    firstName: '',
+                    lastName: '',
+                    username: '',
+                    password: '',
+                    matchingPassword: ''
+                }}
+                    onSubmit={(values) => {
+                        performRegistration(values)
+                            .then(data => console.log(data));
+                    }}
+
+                    validationSchema={Yup.object().shape({
+                        firstName: Yup.string()
+                            .required("Pflichtfeld"),
+                        lastName: Yup.string()
+                            .required("Pflichtfeld"),
+                        username: Yup.string()
+                            .required("Pflichtfeld")
+                            .email("Keine valide E-Mail-Adresse"),
+                        password: Yup.string()
+                            .required("Pflichtfeld")
+                            .min(8, "Min. 8 Zeichen")
+                            .matches(
+                                /^(?=.*[A-Z])(?=.*[a-z])(?=.*[0-9])/,
+                                "Min. ein Groß- sowie Kleinbuchstabe und min. eine Zahl"),
+                        matchingPassword: Yup.string()
+                            .required("Pflichtfeld")
+                            .oneOf([Yup.ref("password"), null], "Passwörter stimmen nicht überein")
+                    })}
+            >
+                {props => {
+                    const {
+                        values,
+                        touched,
+                        errors,
+                        handleChange,
+                        handleBlur,
+                        handleSubmit
+                    } = props;
+                    return (
+                        <Form>
+                            <Grid item className={classes.nextTopic}>
+                                <InputTextFieldValidated fieldType="text" fieldName="firstName"
+                                                         label="Vorname" formikProps={props}/>
+                            </Grid>
+                            <Grid item>
+                                <InputTextFieldValidated fieldType="text" fieldName="lastName"
+                                                         label="Nachname" formikProps={props}/>
+                            </Grid>
+                            <Grid item>
+                                <InputTextFieldValidated fieldType="text" fieldName="username"
+                                                         label="E-Mail-Adresse" formikProps={props}/>
+                            </Grid>
+                            <Grid item>
+                                <InputTextFieldValidated fieldType="password" fieldName="password"
+                                                         label="Passwort" formikProps={props}/>
+                            </Grid>
+                            <Grid item>
+                                <InputTextFieldValidated fieldType="password" fieldName="matchingPassword"
+                                                         label="Passwort wiederholen" formikProps={props}/>
+                            </Grid>
+                            <Grid container
+                                  direction="row"
+                                  justify="space-between">
+                                <Grid item className={classes.nextTopic}>
+                                    <ButtonYellowBig handleClick={props.handleSubmit}
+                                                     buttonText="Account erstellen"/>
+                                </Grid>
+                                <Grid item className={classes.nextTopic}>
+                                    <ButtonYellowBig handleClick={cancel} buttonText="Abbrechen"/>
+                                </Grid>
+                            </Grid>
+                        </Form>
+                    )
+                }}
+            </Formik>
+        </>
+    )
+}
